I've fixed some problems.  I wasn't calling the sfunction before now I am.  Made sure that bits are going into the right places.

+++ DESArrays.java
@@ -1,6 +1,6 @@
 public class DESArrays
 {
-   public static byte[][] getSelectionTables(byte tableNum)
+   public static byte[][] getSelectionTables(int tableNum)
    {
       final byte[][] s1 =
       {
